v0.7.0: fix signing, broadcast, and query bugs
- Fix Cosmos address derivation: use ripemd160(sha256(pubkey)) instead of sha256[..20]
- Fix tx signing: let secp256k1 signer handle hashing internally
- Switch broadcast from LCD to RPC broadcast_tx_sync for reliability
- Handle nested account types (vesting accounts) in get_account_info
- Add simulate error checking before broadcast
- Use lithium-cli 0.1.2 typed structs for all contract query/execute messages
- Fix infra_tx_search: use numeric order_by enum (2=DESC)
- Fix GraphQL messages_by_address: wrap addresses in PostgreSQL array literal
- Change SubmitProofParams.difficulty from u64 to u32

+++ rust/src/tools/mod.rs
@@ -230,11 +230,11 @@ pub struct SubmitProofParams {
     pub nonce: u64,
     #[schemars(description = "Miner address")]
     pub miner_address: String,
-    #[schemars(description = "Challenge string")]
+    #[schemars(description = "Challenge hex string (32 bytes)")]
     pub challenge: String,
-    #[schemars(description = "Difficulty (minimum 1)")]
-    pub difficulty: u64,
-    #[schemars(description = "Timestamp of proof generation")]
+    #[schemars(description = "Difficulty in bits")]
+    pub difficulty: u32,
+    #[schemars(description = "Timestamp (unix seconds)")]
     pub timestamp: u64,
     #[schemars(description = "Optional referrer address")]
     pub referrer: Option<String>,
